requirements.txt
# requirements.txt
langgraph
langchain-ollama
langchain-core
pydantic
python-dotenv
requests
psutil
nvidia-ml-py
fastapi
uvicorn